requirements.txt
assemblyai==0.20.2
pinecone-client==3.0.0
sentence-transformers==2.2.2
